+++ app/src/main/java/org/sopt/dosopttemplate/data/datasource/AuthDataSource.kt
@@ -1,8 +1,10 @@
 package org.sopt.dosopttemplate.data.datasource
 
 import org.sopt.dosopttemplate.data.model.requestModel.RequestLoginDto
+import org.sopt.dosopttemplate.data.model.requestModel.RequestSignupDto
 import org.sopt.dosopttemplate.data.model.responseModel.ResponseLoginDto
 
-interface LoginDataSource {
+interface AuthDataSource {
     suspend fun login(request: RequestLoginDto): ResponseLoginDto
+    suspend fun signup(request: RequestSignupDto)
 }
