+++ app/src/main/java/org/sopt/dosopttemplate/data/datasource/AuthDataSourceImpl.kt
@@ -1,11 +1,16 @@
 package org.sopt.dosopttemplate.data.datasource
 
+import org.sopt.dosopttemplate.data.model.requestModel.RequestLoginDto
 import org.sopt.dosopttemplate.data.model.requestModel.RequestSignupDto
+import org.sopt.dosopttemplate.data.model.responseModel.ResponseLoginDto
 import org.sopt.dosopttemplate.data.service.AuthService
 import javax.inject.Inject
 
-class SignUpDataSourceImpl @Inject constructor(
+class AuthDataSourceImpl @Inject constructor(
     private val authService: AuthService,
-) : SignUpDataSource {
+) : AuthDataSource {
+    override suspend fun login(request: RequestLoginDto): ResponseLoginDto =
+        authService.login(request)
+
     override suspend fun signup(request: RequestSignupDto): Unit = authService.signup(request)
 }
