--- app/src/main/java/org/sopt/dosopttemplate/data/datasource/SignUpDataSource.kt
@@ -1,7 +1,0 @@
-package org.sopt.dosopttemplate.data.datasource
-
-import org.sopt.dosopttemplate.data.model.requestModel.RequestSignupDto
-
-interface SignUpDataSource {
-    suspend fun signup(request: RequestSignupDto)
-}
