+++ app/src/main/java/org/sopt/dosopttemplate/di/BindModule.kt
@@ -4,17 +4,13 @@ import dagger.Binds
 import dagger.Module
 import dagger.hilt.InstallIn
 import dagger.hilt.components.SingletonComponent
-import org.sopt.dosopttemplate.data.datasource.LoginDataSource
-import org.sopt.dosopttemplate.data.datasource.LoginDataSourceImpl
-import org.sopt.dosopttemplate.data.datasource.SignUpDataSource
-import org.sopt.dosopttemplate.data.datasource.SignUpDataSourceImpl
+import org.sopt.dosopttemplate.data.datasource.AuthDataSource
+import org.sopt.dosopttemplate.data.datasource.AuthDataSourceImpl
 import org.sopt.dosopttemplate.data.datasource.UserDataSource
 import org.sopt.dosopttemplate.data.datasource.UserDataSourceImpl
-import org.sopt.dosopttemplate.data.repository.LoginRepositoryImpl
-import org.sopt.dosopttemplate.data.repository.SignUpRepositoryImpl
+import org.sopt.dosopttemplate.data.repository.AuthRepositoryImpl
 import org.sopt.dosopttemplate.data.repository.UserRepositoryImpl
-import org.sopt.dosopttemplate.domain.repository.LoginRepository
-import org.sopt.dosopttemplate.domain.repository.SignUpRepository
+import org.sopt.dosopttemplate.domain.repository.AuthRepository
 import org.sopt.dosopttemplate.domain.repository.UserRepository
 import javax.inject.Singleton
 
@@ -28,11 +24,7 @@ abstract class BindModule {
 
     @Binds
     @Singleton
-    abstract fun bindLoginDataSource(loginDataSourceImpl: LoginDataSourceImpl): LoginDataSource
-
-    @Binds
-    @Singleton
-    abstract fun bindSignupDataSource(signUpDataSourceImpl: SignUpDataSourceImpl): SignUpDataSource
+    abstract fun bindAuthDataSource(AuthDataSourceImpl: AuthDataSourceImpl): AuthDataSource
 
     @Binds
     @Singleton
@@ -40,9 +32,5 @@ abstract class BindModule {
 
     @Binds
     @Singleton
-    abstract fun bindLoginRepository(loginRepositoryImpl: LoginRepositoryImpl): LoginRepository
-
-    @Binds
-    @Singleton
-    abstract fun bindSignupRepository(signupRepositoryImpl: SignUpRepositoryImpl): SignUpRepository
+    abstract fun bindAuthRepository(authRepositoryImpl: AuthRepositoryImpl): AuthRepository
 }
