+++ app/src/main/java/org/sopt/dosopttemplate/domain/repository/AuthRepository.kt
@@ -1,8 +1,10 @@
 package org.sopt.dosopttemplate.domain.repository
 
 import org.sopt.dosopttemplate.data.model.requestModel.RequestLoginDto
+import org.sopt.dosopttemplate.data.model.requestModel.RequestSignupDto
 import org.sopt.dosopttemplate.data.model.responseModel.ResponseLoginDto
 
-interface LoginRepository {
+interface AuthRepository {
     suspend fun login(dto: RequestLoginDto): Result<ResponseLoginDto>
+    suspend fun signup(request: RequestSignupDto): Result<Unit>
 }
