--- app/src/main/java/org/sopt/dosopttemplate/domain/repository/SignUpRepository.kt
@@ -1,7 +1,0 @@
-package org.sopt.dosopttemplate.domain.repository
-
-import org.sopt.dosopttemplate.data.model.requestModel.RequestSignupDto
-
-interface SignUpRepository {
-    suspend fun signup(request: RequestSignupDto): Result<Unit>
-}
