+++ app/src/main/java/org/sopt/dosopttemplate/presentation/auth/login/LoginViewModel.kt
@@ -8,14 +8,14 @@ import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.launch
 import org.sopt.dosopttemplate.data.model.requestModel.RequestLoginDto
 import org.sopt.dosopttemplate.data.model.responseModel.ResponseLoginDto
-import org.sopt.dosopttemplate.domain.repository.LoginRepository
+import org.sopt.dosopttemplate.domain.repository.AuthRepository
 import org.sopt.dosopttemplate.presentation.auth.AuthState
 import org.sopt.dosopttemplate.presentation.auth.signup.SignUpViewModel
 import javax.inject.Inject
 
 @HiltViewModel
 class LoginViewModel @Inject constructor(
-    private val loginRepository: LoginRepository,
+    private val authRepository: AuthRepository,
 ) : ViewModel() {
     val isMoveSignupActivity: MutableLiveData<Boolean> = MutableLiveData(false)
 
@@ -49,7 +49,7 @@ class LoginViewModel @Inject constructor(
 
     fun clickLoginBtn() {
         viewModelScope.launch {
-            loginRepository.login(RequestLoginDto(id.value ?: "", pw.value ?: "")).onSuccess {
+            authRepository.login(RequestLoginDto(id.value ?: "", pw.value ?: "")).onSuccess {
                 _loginResult.value = it
                 _loginSuccess.value = true
             }.onFailure {
